Do not remove the database if there is no backup and deal with manifest
In pg13, when creating a backup of the database, the pg_basebackup utility
will create a file called backup_manifest in a temporary file. smdba does
not take this into account, and when you run smdba backup-hot a second time,
smdba aborts because it finds this file already.

The issue is that smdba aborts the execution after having removed the database,
not before, so there is no backup to be restored.

This commit deals with backup_manifest file, and abort a restore if there
is no backup to be restored, before removing the actual database.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>